Add files via upload
A cash-flow diagram is a financial tool used to represent the cashflows associated with a security, "project", or business.

As per the graphics, cash flow diagrams are widely used in structuring and analyzing securities, particularly swaps. They may also be used to represent payment schedules for bonds, mortgages and other types of loans.